the first 50 Fibonacci numbers

+++ 0x02-functions_nested_loops/102-fibonacci.c
@@ -8,14 +8,14 @@
 int main(void)
 {
 	int i = 3;
-	unsigned long long int f = 1, s = 2, next;
+	unsigned long f = 1, s = 2, next;
 
-	printf("%llu, ", f);
-	printf(", %llu", s);
+	printf("%lu, ", f);
+	printf(", %lu", s);
 	while (i <= 50)
 	{
 		next = f + s;
-		 printf(", %llu", next);
+		 printf(", %lu", next);
 		f = s;
 		s = next;
 		i++;
